Refactor admin into a module with separate files
This was getting somewhat unmanageable
